[refactor]: changed icon pathname before: ?react.svg after: .svg
