+++ src/icons/icon/SelectedCheckboxIcon.tsx
@@ -1,4 +1,4 @@
-import SelectedCheckboxSvg from '@/assets/icon/selected_checkbox.svg?react';
+import SelectedCheckboxSvg from '@/assets/icon/selected_checkbox.svg';
 import { IconProps } from '../types';
 
 const SelectedCheckboxIcon = ({ width, height, className }: IconProps) => {
